Созданы классы (TestBase, Test_CreateContact, Test_CreateGroup, диаграмма классов)
