+++ addressbook-web-tests/AccountData.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace addressbook_web_tests
 {
-    class AccountData
+    public class AccountData : TestBase
     {
         private string username;
         private string password;
